feat: added first iteration of butterworth 4th order filter using arm bicascade function

+++ AcousticsMicrocontrollerCode/teensy41/src/ethernet_module.cpp
@@ -1,4 +1,5 @@
 #include "ethernet_module.h"
+#include <cstddef>
 
 
 byte macAddressTeensy[] = {0x00, 0x01, 0xBE, 0xEE, 0xFE, 0xED};
@@ -86,7 +87,7 @@ char* UDP_read_message() {
 }
 
 
-void UDP_send_message_raw(void* data, uint32_t dataSize){
+void UDP_send_message_raw(void* data, size_t dataSize){
     IPAddress remoteIP = Udp.remoteIP();
     uint16_t remotePort = Udp.remotePort();
